+++ src/command/clanbattle.ts
@@ -2,18 +2,19 @@ import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
 import {NtoA} from 'alphabet-to-number'
-import * as command from './'
-import * as util from '../../util'
-import * as schedule from '../../io/schedule'
-import * as status from '../../io/status'
-import {AtoE} from '../../io/type'
-import * as etc from '../convex/etc'
-import * as format from '../convex/format'
-import * as lapAndBoss from '../convex/lapAndBoss'
-import * as manage from '../convex/manage'
-import * as situation from '../convex/situation'
-import * as declare from '../declare/list'
-import * as list from '../plan/list'
+import * as command from '.'
+import * as util from '../util'
+import * as schedule from '../io/schedule'
+import * as status from '../io/status'
+import {AtoE} from '../io/type'
+import * as etc from '../client/convex/etc'
+import * as role from '../client/convex/role'
+import * as format from '../client/convex/format'
+import * as lapAndBoss from '../client/convex/lapAndBoss'
+import * as manage from '../client/convex/manage'
+import * as situation from '../client/convex/situation'
+import * as declare from '../client/declare/list'
+import * as list from '../client/plan/list'
 
 /**
  * クラバト用のコマンド
@@ -303,7 +304,7 @@ const overController = async (_command: string, _content: string, _msg: Discord.
  */
 const taskKillController = async (_command: string, _content: string, _msg: Discord.Message) => {
   // タスキルのロールを付与する
-  etc.AddTaskKillRole(_msg)
+  role.AddTaskKillRole(_msg)
 }
 
 /**
