+++ src/command/management.ts
@@ -1,18 +1,18 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
-import * as command from './'
-import * as util from '../../util'
-import * as bossTable from '../../io/bossTable'
-import * as dateTable from '../../io/dateTable'
-import * as schedule from '../../io/schedule'
-import * as status from '../../io/status'
-import * as category from '../convex/category'
-import * as etc from '../convex/etc'
-import * as react from '../convex/react'
-import * as situation from '../convex/situation'
-import * as list from '../plan/list'
-import * as plan from '../plan/delete'
+import * as command from '.'
+import * as util from '../util'
+import * as bossTable from '../io/bossTable'
+import * as dateTable from '../io/dateTable'
+import * as schedule from '../io/schedule'
+import * as status from '../io/status'
+import * as category from '../client/convex/etc/category'
+import * as etc from '../client/convex/etc'
+import * as react from '../client/convex/react'
+import * as situation from '../client/convex/situation'
+import * as list from '../client/plan/list'
+import * as plan from '../client/plan/delete'
 
 /**
  * 運営管理者用のコマンド
