+++ src/convex/manage.ts
@@ -3,8 +3,8 @@ import Settings from 'const-settings'
 import * as util from '../../util'
 import * as status from '../../io/status'
 import {Member} from '../../io/type'
-import * as etc from './etc'
-import * as time from './time'
+import * as role from './role'
+import * as time from './time/time'
 import * as situation from './situation'
 
 /**
@@ -57,7 +57,7 @@ const convexEndProcess = async (member: Member, user: Discord.User, msg: Discord
   // ロールを削除
   const guildMember = await util.MemberFromId(user.id)
   await guildMember.roles.remove(Settings.ROLE_ID.REMAIN_CONVEX)
-  await etc.RemoveBossRole(guildMember)
+  await role.RemoveBossRole(guildMember)
 
   // 何人目の3凸終了者なのかを報告する
   const members = await status.Fetch()
