+++ src/convex/time/limit.ts
@@ -1,10 +1,10 @@
 import * as Discord from 'discord.js'
 import Option from 'type-of-option'
 import Settings from 'const-settings'
-import * as dateTable from '../../io/dateTable'
-import * as status from '../../io/status'
-import * as util from '../../util'
-import {Member} from '../../io/type'
+import * as dateTable from '../../../io/dateTable'
+import * as status from '../../../io/status'
+import * as util from '../../../util'
+import {Member} from '../../../io/type'
 
 /**
  * 活動限界時間を設定する
