+++ src/index.ts
@@ -3,7 +3,7 @@ import * as Discord from 'discord.js'
 import ThrowEnv from 'throw-env'
 import {Partials, Intents} from './init/setting'
 import {ClientEvent} from './init/event'
-import {CronOperation} from './init/cron'
+import {CronOperation} from './cron'
 
 export const Client = new Discord.Client({
   partials: Partials,
