+++ tsconfig.json
@@ -15,55 +15,56 @@
     "outDir": "./dist/"
   },
   "files": [
+    "src/cron.ts",
+    "src/event.ts",
     "src/index.ts",
-      "src/client/guildMemberAdd.ts",
-      "src/client/guildMemberUpdate.ts",
-      "src/client/messageCreate.ts",
-      "src/client/messageDelete.ts",
-      "src/client/messageReactionAdd.ts",
-      "src/client/messageReactionRemove.ts",
-      "src/client/messageUpdate.ts",
-      "src/client/ready.ts",
-        "src/client/command/clanbattle.ts",
-        "src/client/command/index.ts",
-        "src/client/command/management.ts",
-        "src/client/convex/attendance.ts",
-        "src/client/convex/category.ts",
-        "src/client/convex/etc.ts",
-        "src/client/convex/format.ts",
-        "src/client/convex/lapAndBoss.ts",
-        "src/client/convex/manage.ts",
-        "src/client/convex/over.ts",
-        "src/client/convex/react.ts",
-        "src/client/convex/sister.ts",
-        "src/client/convex/situation.ts",
-        "src/client/convex/time.ts",
-        "src/client/declare/list.ts",
-        "src/client/declare/index.ts",
-        "src/client/declare/react.ts",
-        "src/client/declare/status.ts",
-        "src/client/etc/emoji.ts",
-        "src/client/etc/playerID.ts",
-        "src/client/etc/send.ts",
-        "src/client/plan/delete.ts",
-        "src/client/plan/edit.ts",
-        "src/client/plan/index.ts",
-        "src/client/plan/list.ts",
-        "src/client/plan/update.ts",
-        "src/client/report/cancel.ts",
-        "src/client/report/index.ts",
-        "src/client/report/update.ts",
-      "src/io/bossTable.ts",
-      "src/io/current.ts",
-      "src/io/dateTable.ts",
-      "src/io/index.ts",
-      "src/io/json.ts",
-      "src/io/schedule.ts",
-      "src/io/status.ts",
-      "src/io/type.ts",
-      "src/util/cron.ts",
-      "src/util/index.ts",
-      "src/util/prototype.ts",
-      "src/init/setting.ts",
+    "src/setting.ts",
+    "src/client/guildMemberAdd.ts",
+    "src/client/guildMemberUpdate.ts",
+    "src/client/messageCreate.ts",
+    "src/client/messageDelete.ts",
+    "src/client/messageReactionAdd.ts",
+    "src/client/messageReactionRemove.ts",
+    "src/client/messageUpdate.ts",
+    "src/client/ready.ts",
+    "src/command/clanbattle.ts",
+    "src/command/index.ts",
+    "src/command/management.ts",
+    "src/convex/lapAndBoss.ts",
+    "src/convex/manage.ts",
+    "src/convex/over.ts",
+    "src/convex/react.ts",
+    "src/convex/role.ts",
+    "src/convex/situation.ts",
+    "src/convex/declare/index.ts",
+    "src/convex/declare/list.ts",
+    "src/convex/declare/react.ts",
+    "src/convex/declare/status.ts",
+    "src/convex/etc/category.ts",
+    "src/convex/etc/format.ts",
+    "src/convex/etc/index.ts",
+    "src/convex/plan/delete.ts",
+    "src/convex/plan/edit.ts",
+    "src/convex/plan/index.ts",
+    "src/convex/plan/list.ts",
+    "src/convex/plan/update.ts",
+    "src/convex/report/cancel.ts",
+    "src/convex/report/index.ts",
+    "src/convex/report/update.ts",
+    "src/convex/time/attendance.ts",
+    "src/convex/time/limit.ts",
+    "src/io/bossTable.ts",
+    "src/io/current.ts",
+    "src/io/dateTable.ts",
+    "src/io/json.ts",
+    "src/io/redis.ts",
+    "src/io/schedule.ts",
+    "src/io/status.ts",
+    "src/util/emoji.ts",
+    "src/util/index.ts",
+    "src/util/playerID.ts",
+    "src/util/prototype.ts",
+    "src/util/send.ts",
+    "src/util/type.ts",
   ]
 }
